+++ src/drom_lib/main.ml
@@ -34,6 +34,7 @@ let main () =
       CommandPromote.cmd;
       CommandLock.cmd;
       CommandDep.cmd;
+      CommandOpamPlugin.cmd;
     ]
   in
   let common_args =
@@ -68,7 +69,8 @@ let main () =
         exit 0
     | _ -> args
   in
-  let args = iter_initial_args (List.tl args) in
+
+  let args = iter_initial_args (List.tl args ) in
   let argv = Array.of_list ( Sys.argv.(0) :: args ) in
   (* OpambinMisc.global_log "args: %s"
          (String.concat " " (Array.to_list Sys.argv)); *)