+++ src/drom_lib/opam.ml
@@ -179,14 +179,8 @@ let exec ?(y = false) cmd args =
            [] )
        @ args ))
 
-let root = lazy (
-  try Sys.getenv "OPAMROOT"
-  with Not_found -> Globals.home_dir // ".opam"
-)
-let root () = Lazy.force root
-
 let init ?y ?switch ?edition () =
-  let opam_root = root () in
+  let opam_root = Globals.opam_root () in
 
   if not (Sys.file_exists opam_root) then
     let args =